Modify p1152.cpp
s.length() to (int)s.length()

+++ p1152.cpp
@@ -9,9 +9,9 @@ int		main(void)
 	int		ans = 0;
 
 	getline(cin, s);
-	for (int i = 0; i < s.length() - 1; i++) {
+	for (int i = 0; i < (int)s.length() - 1; i++) {
 		if (s[i] != ' ' && s[i + 1] == ' ') ans++;
-		else if (i + 1 == s.length() - 1 && s[i + 1] != ' ') ans++;
+		else if (i + 1 == (int)s.length() - 1 && s[i + 1] != ' ') ans++;
 	}
 	cout << ans;
 	return (0);
